# backend/api-docs.json
{"openapi":"3.1.0","info":{"title":"API 서버","version":"v1"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/menu/{menu_id}":{"get":{"tags":["menu-controller"],"operationId":"getMenuById","parameters":[{"name":"menu_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RsDataEmpty"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Menu"}}}}}},"put":{"tags":["menu-controller"],"operationId":"updateMenu","parameters":[{"name":"menu_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Menu"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RsDataEmpty"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RsDataMenuDto"}}}}}},"delete":{"tags":["menu-controller"],"operationId":"deleteMenu","parameters":[{"name":"menu_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RsDataEmpty"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RsDataVoid"}}}}}}},"/menu/{id}/edit":{"put":{"tags":["home-controller"],"operationId":"modify","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"menu","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Menu"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RsDataEmpty"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/order":{"get":{"tags":["order-controller"],"summary":"email로 주문 조회","operationId":"getOrdersByEmail","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RsDataEmpty"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OrderMenuWithOrderDto"}}}}}}},"post":{"tags":["order-controller"],"summary":"주문 생성","operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderReqBody"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RsDataEmpty"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RsDataOrderDto"}}}}}}},"/menu":{"get":{"tags":["menu-controller"],"operationId":"getAllMenus","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RsDataEmpty"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RsDataListMenuDto"}}}}}},"post":{"tags":["menu-controller"],"operationId":"addMenu","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Menu"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RsDataEmpty"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RsDataMenuDto"}}}}}}},"/menu/add":{"get":{"tags":["home-controller"],"operationId":"add","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RsDataEmpty"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}},"post":{"tags":["home-controller"],"operationId":"add_1","parameters":[{"name":"menu","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Menu"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RsDataEmpty"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/order-menu":{"get":{"tags":["admin-controller"],"summary":"전체 주문 조회","operationId":"getAllOrders","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RsDataEmpty"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RsDataListOrderMenuDto"}}}}}}},"/menu_manage":{"get":{"tags":["home-controller"],"operationId":"menu_list","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RsDataEmpty"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/menu/{id}/modify":{"get":{"tags":["home-controller"],"operationId":"showModifyPage","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RsDataEmpty"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/menu/{id}/delete":{"get":{"tags":["home-controller"],"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RsDataEmpty"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/list":{"get":{"tags":["home-controller"],"operationId":"list","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"kw","in":"query","required":false,"schema":{"type":"string","default":""}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RsDataEmpty"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/detail/{id}":{"get":{"tags":["home-controller"],"operationId":"detail","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RsDataEmpty"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"Empty":{"type":"object"},"RsDataEmpty":{"type":"object","properties":{"resultCode":{"type":"string"},"msg":{"type":"string"},"data":{"$ref":"#/components/schemas/Empty"}},"required":["data","msg","resultCode"]},"Menu":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"type":{"type":"string"},"price":{"type":"integer","format":"int32"}}},"MenuDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"type":{"type":"string"},"price":{"type":"integer","format":"int32"}}},"RsDataMenuDto":{"type":"object","properties":{"resultCode":{"type":"string"},"msg":{"type":"string"},"data":{"$ref":"#/components/schemas/MenuDto"}},"required":["data","msg","resultCode"]},"OrderReqBody":{"type":"object","properties":{"email":{"type":"string"},"address":{"type":"string"},"postalCode":{"type":"integer","format":"int32","minimum":1},"orders":{"type":"object","additionalProperties":{"type":"integer","format":"int32"}}},"required":["address","orders","postalCode"]},"OrderDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"type":"string"},"address":{"type":"string"},"postalCode":{"type":"integer","format":"int32"},"orders":{"type":"object","additionalProperties":{"type":"integer","format":"int32"}}}},"RsDataOrderDto":{"type":"object","properties":{"resultCode":{"type":"string"},"msg":{"type":"string"},"data":{"$ref":"#/components/schemas/OrderDto"}},"required":["data","msg","resultCode"]},"MenuDataDto":{"type":"object","properties":{"menuId":{"type":"integer","format":"int64"},"menuName":{"type":"string"},"menuPrice":{"type":"integer","format":"int32"},"menuType":{"type":"string"},"menuCount":{"type":"integer","format":"int32"}}},"OrderMenuWithOrderDto":{"type":"object","properties":{"orderId":{"type":"integer","format":"int64"},"email":{"type":"string"},"address":{"type":"string"},"postalCode":{"type":"integer","format":"int32"},"menuData":{"type":"array","items":{"$ref":"#/components/schemas/MenuDataDto"}},"orderTime":{"type":"string","format":"date-time"},"totalPrice":{"type":"integer","format":"int32"},"after2pm":{"type":"boolean"}}},"OrderMenuDto":{"type":"object","properties":{"orderId":{"type":"integer","format":"int64"},"menuData":{"type":"array","items":{"$ref":"#/components/schemas/MenuDataDto"}},"orderTime":{"type":"string","format":"date-time"},"totalPrice":{"type":"integer","format":"int32"},"after2pm":{"type":"boolean"}}},"RsDataListOrderMenuDto":{"type":"object","properties":{"resultCode":{"type":"string"},"msg":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/OrderMenuDto"}}},"required":["data","msg","resultCode"]},"RsDataListMenuDto":{"type":"object","properties":{"resultCode":{"type":"string"},"msg":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/MenuDto"}}},"required":["data","msg","resultCode"]},"RsDataVoid":{"type":"object","properties":{"resultCode":{"type":"string"},"msg":{"type":"string"},"data":{"type":"object"}},"required":["data","msg","resultCode"]}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}